requirements.txt
fastapi
uvicorn
pydantic
pydantic[email]
asyncpg
python-dotenv
supabase
argon2-cffi
python-jose[cryptography]
